chore(api): refactor to single address and code cleanup
BREAKING CHANGE: Dropping `LoadAllArtifacts` and `LoadArtifacts` and wallets/keys/keystore related functions from `api.User`.
`api.Sign` won't accept anymore a pub key as parameter.
